API modifications #2

+++ api.php
@@ -1,4 +1,6 @@
 <?php
+if (!empty($argv[0])) chdir(pathinfo($argv[0], PATHINFO_DIRNAME));
+
 require ('config.php');
 require (CLASSES.'event.php');
 require (FUNCTIONS.'functions.php');
@@ -24,14 +26,16 @@ if (isset($_GET['playlist'])) {
 } elseif (isset($_GET['maintenance']) or $argv[1] == 'maintenance') {
     $files = scanFolder(DATA, ['media', '.', '..']);
     $count = 0;
-    foreach ($files as $file) {
-        $ev = new Event();
-        if ($ev->read($file)) {
-            if (isset($ev->event['retention_time']) and $ev->event['retention_time'] < time()) {
-                if (is_file(MEDIA.basename($ev->event['icon']))) unlink(MEDIA.basename($ev->event['icon']));
-                if (is_file(MEDIA.basename($ev->event['fanart']))) unlink(MEDIA.basename($ev->event['fanart']));
-                unlink(DATA . $file);
-                $count++;
+    if (!empty($files)) {
+        foreach ($files as $file) {
+            $ev = new Event();
+            if ($ev->read($file)) {
+                if (isset($ev->event['retention_time']) and $ev->event['retention_time'] < time()) {
+                    if (is_file(MEDIA.basename($ev->event['icon']))) unlink(MEDIA.basename($ev->event['icon']));
+                    if (is_file(MEDIA.basename($ev->event['fanart']))) unlink(MEDIA.basename($ev->event['fanart']));
+                    unlink(DATA . $file);
+                    $count++;
+                }
             }
         }
     }
